move to Internal\CodeGen + fix build for sln
