+++ src/protobuf-net.BuildTools/Internal/CodeGen/Parsers/Common/TypeCodeGenModelParserBase.cs
@@ -1,10 +1,10 @@
 ﻿using System.Collections.Immutable;
 using Microsoft.CodeAnalysis;
 using ProtoBuf.BuildTools.Internal;
-using ProtoBuf.Internal.CodeGenSemantic.Abstractions;
-using ProtoBuf.Internal.CodeGenSemantic.Models;
+using ProtoBuf.Internal.CodeGen.Abstractions;
+using ProtoBuf.Internal.CodeGen.Models;
 
-namespace ProtoBuf.Internal.CodeGenSemantic.Parsers.Common;
+namespace ProtoBuf.Internal.CodeGen.Parsers.Common;
 
 internal abstract class TypeCodeGenModelParserBase<TCodeGenModel> : ISymbolCodeGenModelParser<ITypeSymbol, TCodeGenModel>
 {
