+++ src/protobuf-net.BuildTools/Internal/CodeGen/Parsers/EnumPropertyCodeGenModelParser.cs
@@ -1,12 +1,12 @@
 ﻿using System.Collections.Immutable;
 using Microsoft.CodeAnalysis;
 using ProtoBuf.BuildTools.Internal;
-using ProtoBuf.Internal.CodeGenSemantic.Abstractions;
-using ProtoBuf.Internal.CodeGenSemantic.Models;
-using ProtoBuf.Internal.CodeGenSemantic.Parsers.Common;
+using ProtoBuf.Internal.CodeGen.Abstractions;
+using ProtoBuf.Internal.CodeGen.Models;
+using ProtoBuf.Internal.CodeGen.Parsers.Common;
 using ProtoBuf.Reflection.Internal.CodeGen;
 
-namespace ProtoBuf.Internal.CodeGenSemantic.Parsers;
+namespace ProtoBuf.Internal.CodeGen.Parsers;
 
 internal sealed class EnumPropertyCodeGenModelParser : PropertyCodeGenModelParserBase<CodeGenEnum>
 {
